[ncp] implement ncp version of otPlatDnssd (#10882)
This commit implements NCP based version of `otPlatDnssdGetState`.

A unit test is added to verify the implementation. This commit also
adjusts the unit test workflow to put the ncp unit tests into a
separate item. The direct reason for this commit is that `otPlatDnssd`
APIs haven't been implemented in `FakePlatform` and when
`OT_PLATFORM_DNSSD` is turned, the build failed. Let's first run NCP
unit tests separately and later try to merge the test items.