[border-agent] check that BA is enabled in `RegisterService()` (#11993)
This change adds a check at the beginning of `RegisterService()`
to ensure that the border agent is enabled before attempting to
register a service.

Previously, methods like `SetServiceBaseName()` would call
`RegisterService()` regardless of whether the agent was enabled,
leading to unintended service registration attempts on a disabled
agent.

This commit also adds a new test case to verify that changing the
service base name on a disabled border agent does not result in a
service being registered.