[border-router] make `RxRaTracker` handle netdata events directly (#12016)
This change makes the `RxRaTracker` a direct listener of network data
change events from the `Notifier`.

Previously, the `RoutingManager` would receive the network data change
event and then call `RxRaTracker::HandleNetDataChange()`. This
created an unnecessary dependency between the two components.

By making `RxRaTracker` a direct listener, we decouple it from
`RoutingManager`. The `HandleNetDataChange()` method in `RxRaTracker`
is also made private as it is now only called from within the class.