[posix] improve infra interface state detection (#12194)
This commit enhances the robustness of interface management with two
improvements:

1. The `SendIcmp6Nd` function now explicitly handles `EADDRNOTAVAIL`
and `ENODEV` errors from the `sendmsg()` system call. These errors
often indicate that the interface is down or has lost its
addresses. Upon that, we trigger an immediate re-evaluation of the
interface's state by calling `otPlatInfraIfStateChanged()`.

2. The `GetFlags` function has been updated to also verify mInfraIndex
and mInfraName not changed.